Extract the app-scoped object graph out of App.kt
App.kt was 1266 lines holding four unrelated things: the object graph,
the navigation graph, startup routing, and the favorites-review domain.
The graph was the worst of them — twenty `remember` calls between the
imports and the first screen, so reading how a destination is wired meant
scrolling past every service the app owns.

Three moves, all verbatim:

  - AppContainer.kt — the graph, built by rememberAppContainer(). Each
    object keeps the exact `remember` keys it had, so a service is still
    recreated when its own inputs change and not when a sibling's do.
    Collapsing them into one keyed remember would have tied, for example,
    the app database's lifetime to androidContext, which it never had.
    The two objects that own resources are closed there too, next to
    where they are made, rather than in an effect further down App.kt.

  - data/learning/review/FavoritesReviewCoordinator.kt — the coordinator
    with FavoritesReviewState and FavoriteLanguageReviewState. It reads
    favorites, runs intake and computes queue stats; none of that is UI
    entry-point work. Its test moves with it.

  - AppDestination.kt — the destinations, now internal rather than
    private since they are read from App.kt.

toFavoriteLanguageReviewUiState stays in App.kt: it maps to a UI type and
belongs on the UI side of the boundary, not in data.

The container exposes fourteen members, not eighteen. appDatabase,
listsClient, wordListsRepository and fsrsConfig exist only to build the
others, and nothing outside the graph asks for them, so they stay local
to rememberAppContainer. Publishing them would invite a screen to reach
past a service to what it was built from.

ViewModel lifetimes are unchanged, which is what this refactor most
risked: ten screens still use viewModel(viewModelStoreOwner =
backStackEntry), and the app-scoped Favorites and Settings ViewModels
plus the bounded Word Details cache still use remember, exactly as
before. Counts verified against the previous revision.

App.kt is 967 lines. Verified mechanically that the 82 lines of graph,
the coordinator, and the destinations moved byte for byte, the
private-to-internal change on AppDestination aside. Twenty-two imports
left behind by the moves are gone.

CLAUDE.md described App.kt as where services are created; updated to
match, including the rule about not growing the container speculatively.

Verified: :composeApp:compileKotlinDesktop, :composeApp:verifyLocalizationKeys,
:composeApp:desktopTest (465 tests, the same 5 DictionaryClientTest
failures that need live GitHub and AI backends).
